+++ app/messaging/send-crm-message.js
@@ -1,6 +1,6 @@
 const config = require('../config')
 
-const { CRM: CRM_MESSAGE_TYPE } = require('../constants/message-type')
+const { CRM: CRM_MESSAGE_TYPE } = require('../constants/message-types')
 const SOURCE = require('../constants/message-source')
 
 const { getMessage } = require('../processing/crm')
